+++ plugin_ide.core/src-lang/melnorme/lang/ide/core/utils/prefs/IPrefChangeListener.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2014, 2014 Bruno Medeiros and other Contributors.
+ * Copyright (c) 2014, 2014 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -8,10 +8,10 @@
  * Contributors:
  *     Bruno Medeiros - initial API and implementation
  *******************************************************************************/
-package melnorme.util.swt.components;
+package melnorme.lang.ide.core.utils.prefs;
 
-public interface IDisposable {
+public interface IPrefChangeListener {
 	
-	void dispose();
+	void handleChange();
 	
 }