+++ plugin_ide.core/src-lang/melnorme/lang/ide/core/utils/prefs/PreferenceHelper.java
@@ -1,0 +1,90 @@
+/*******************************************************************************
+ * Copyright (c) 2014, 2014 Bruno Medeiros and other Contributors.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ *
+ * Contributors:
+ *     Bruno Medeiros - initial API and implementation
+ *******************************************************************************/
+package melnorme.lang.ide.core.utils.prefs;
+
+import static melnorme.utilbox.core.Assert.AssertNamespace.assertTrue;
+
+import java.util.HashMap;
+
+import melnorme.lang.ide.core.LangCore;
+
+import org.eclipse.core.runtime.preferences.DefaultScope;
+import org.eclipse.core.runtime.preferences.IEclipsePreferences;
+
+public abstract class PreferenceHelper<T> {
+	
+	protected static final HashMap<String, PreferenceHelper<?>> instances = new HashMap<>();
+	
+	public final String pluginId;
+	public final String key;
+	protected final T defaultValue;
+	
+	public PreferenceHelper(String key, T defaultValue) {
+		this(LangCore.PLUGIN_ID, key, defaultValue);
+	}
+	
+	public PreferenceHelper(String pluginId, String key, T defaultValue) {
+		this.pluginId = pluginId;
+		this.key = key;
+		this.defaultValue = defaultValue;
+		
+		initializeDefaultValueInDefaultScope();
+		
+		synchronized (instances) {
+			// Allow only one instance of a preference helper per key.
+			assertTrue(instances.containsKey(key) == false);
+			instances.put(key, this);
+		}
+	}
+	
+	public final String getQualifier() {
+		return pluginId;
+	}
+	
+	protected PreferencesLookupHelper prefs() {
+		return new PreferencesLookupHelper(pluginId);
+	}
+	
+	public abstract T get();
+	
+	protected abstract void initializeDefaultValueInDefaultScope();
+	
+	protected IEclipsePreferences getDefaultPreferences() {
+		return DefaultScope.INSTANCE.getNode(getQualifier());
+	}
+	
+//	public IPreferenceChangeListener_Ext addPrefChangeListener(final IPrefChangeListener listener) {
+//		final IEclipsePreferences node = InstanceScope.INSTANCE.getNode(getQualifier());
+//		IPreferenceChangeListener_Ext prefListener = new IPreferenceChangeListener_Ext() {
+//			
+//			@Override
+//			public void preferenceChange(PreferenceChangeEvent event) {
+//				if(event.getKey().equals(key)) {
+//					assertTrue(event.getNewValue().equals(get()));
+//					
+//					listener.handleChange();
+//				}
+//			}
+//			
+//			@Override
+//			public void dispose() {
+//				node.removePreferenceChangeListener(this);
+//			}
+//		};
+//		node.addPreferenceChangeListener(prefListener);
+//		return prefListener;
+//	}
+//	
+//	public static interface IPreferenceChangeListener_Ext extends IPreferenceChangeListener, IDisposable {
+//		
+//	}
+	
+}