+++ documentation/ChangeLog.md
@@ -7,8 +7,9 @@
   * However, there is a small change using GoClipse with a version other than the recommned/tested one will break things. Each GoClipse release will list a recommented/tested CDT version.
  * Updated/tested GoClipse to CDT 4.4. 
   * Added Dynamic printf ruler action to source editor.
-  * For more info on these new CDT 4.4 debug features, see: https://wiki.eclipse.org/CDT/User/NewIn84#Debug 
- 
+  * For more info on these new CDT 4.4 debug features, see: https://wiki.eclipse.org/CDT/User/NewIn84#Debug
+ * Added build console preference page. Can now customize build console text colors. 
+
 
 ### Goclipse 0.8.0
  * Added debugging support using CDT's integration with GDB.